seperate the g_vbll and vbll
