+++ src/probly/method/__init__.py
@@ -23,6 +23,7 @@ from probly.method.efficient_credal_prediction import efficient_credal_predictio
 from probly.method.ensemble import ensemble
 from probly.method.evidential.classification import evidential_classification
 from probly.method.evidential.regression import evidential_regression
+from probly.method.g_vbll import g_vbll
 from probly.method.graph_posterior_network import (
     cuq_graph_neural_network,
     graph_posterior_network,
@@ -32,7 +33,7 @@ from probly.method.het_net import het_net
 import probly.method.laplace  # noqa: F401  # registers LaplaceRepresenter lazily
 from probly.method.sngp import sngp
 from probly.method.subensemble import subensemble
-from probly.method.vbll import g_vbll, vbll
+from probly.method.vbll import vbll
 
 __all__ = [
     "batchensemble",
