+++ src/probly/method/g_vbll/__init__.py
@@ -1,0 +1,23 @@
+"""G-VBLL: Generative Variational Bayesian Last Layers implementation."""
+
+from __future__ import annotations
+
+from probly.lazy_types import TORCH_MODULE
+
+from ._common import (
+    GVBLLPredictor,
+    g_vbll,
+    g_vbll_traverser,
+)
+
+
+## Torch
+@g_vbll_traverser.delayed_register(TORCH_MODULE)
+def _(_: type) -> None:
+    from . import torch as torch  # noqa: PLC0415
+
+
+__all__ = [
+    "GVBLLPredictor",
+    "g_vbll",
+]
