+++ src/probly/method/vbll/__init__.py
@@ -5,12 +5,9 @@ from __future__ import annotations
 from probly.lazy_types import TORCH_MODULE, TORCH_TENSOR, TORCH_TENSOR_LIKE
 
 from ._common import (
-    GVBLLPredictor,
     VBLLPredictor,
     VBLLRepresenter,
     compute_vbll_categorical_sample,
-    g_vbll,
-    g_vbll_traverser,
     vbll,
     vbll_traverser,
 )
@@ -18,7 +15,6 @@ from ._common import (
 
 ## Torch
 @vbll_traverser.delayed_register(TORCH_MODULE)
-@g_vbll_traverser.delayed_register(TORCH_MODULE)
 def _(_: type) -> None:
     from . import torch as torch  # noqa: PLC0415
 
@@ -29,9 +25,7 @@ def _(_: type) -> None:
 
 
 __all__ = [
-    "GVBLLPredictor",
     "VBLLPredictor",
     "VBLLRepresenter",
-    "g_vbll",
     "vbll",
 ]
