+++ tests/probly/method/g_vbll/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for G-VBLL (Generative Variational Bayesian Last Layers) methods."""
